requirements.txt
azure-identity
azure-mgmt-resource
azure-cosmos
pandas
openpyxl
pyodbc